Moved class functions to class scope.
